requirements.txt
#  Database
psycopg2-binary==2.9.5

#  Enviroment variables
python-dotenv==0.21.0

#  GUI desinger
tkdesigner==1.0.6

#  Util (optional)
alive-progress==3.0.1

#  Time
pytz==